# gen/api/dmss_api/models/document_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from dmss_api.models.base_model_ import Model
from dmss_api import util


class DocumentResponse(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, document=None):  # noqa: E501
        """DocumentResponse - a model defined in OpenAPI

        :param document: The document of this DocumentResponse.  # noqa: E501
        :type document: object
        """
        self.openapi_types = {
            'document': object
        }

        self.attribute_map = {
            'document': 'document'
        }

        self._document = document

    @classmethod
    def from_dict(cls, dikt) -> 'DocumentResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The DocumentResponse of this DocumentResponse.  # noqa: E501
        :rtype: DocumentResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def document(self):
        """Gets the document of this DocumentResponse.


        :return: The document of this DocumentResponse.
        :rtype: object
        """
        return self._document

    @document.setter
    def document(self, document):
        """Sets the document of this DocumentResponse.


        :param document: The document of this DocumentResponse.
        :type document: object
        """

        self._document = document