# gen/api/dmss_api/openapi/openapi.yaml
openapi: 3.0.1
info:
  description: Data storage service for DMT
  title: Data Modelling Storage Service API
  version: 0.1.0
servers:
- url: http://0.0.0.0:8000/api/v1
...
paths:
...
  /explorer/{dataSourceId}/add-package:
    post:
      operationId: add_package
      parameters:
      - description: The data source ID
        explode: false
        in: path
        name: dataSourceId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/DocumentBody'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
          description: Added package
      summary: Add package
      tags:
      - explorer
      x-openapi-router-controller: controllers.explorer_controller
  /explorer/{dataSourceId}/add-raw:
    post:
      operationId: add_raw
      parameters:
      - description: The data source ID
        explode: false
        in: path
        name: dataSourceId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/DocumentBody'
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: Added document UID
      summary: Add raw document
      tags:
      - explorer
      x-openapi-router-controller: controllers.explorer_controller
  /explorer/{dataSourceId}/add-to-parent:
    post:
      operationId: add_to_parent
      parameters:
      - description: The data source ID
        explode: false
        in: path
        name: dataSourceId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/inline_object'
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                parentId:
                  type: string
                type:
                  type: string
                attribute:
                  type: string
              required:
              - attribute
              - name
              - parentId
              - type
              type: object
        description: A JSON object containing document information
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
          description: The UID of the added document
      summary: Add document to parent
      tags:
      - explorer
      x-openapi-router-controller: controllers.explorer_controller
  /explorer/{dataSourceId}/add-to-path:
    post:
      operationId: add_to_path
      parameters:
      - description: The data source ID
        explode: false
        in: path
        name: dataSourceId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/DocumentBody'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
          description: Added document
      summary: Add document to path
      tags:
      - explorer
      x-openapi-router-controller: controllers.explorer_controller
...
components:
  requestBodies:
    DocumentBody:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DocumentBody'
      description: Object containing all info for a document
      required: true
    inline_object:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_object'
      required: true
...
  schemas:
    DocumentResponse:
      additionalProperties:
        type: object
      type: object
    DocumentBody:
      additionalProperties: true
      type: object
...
    inline_object:
      properties:
        name:
          type: string
        parentId:
          type: string
        type:
          type: string
        attribute:
          type: string
      required:
      - attribute
      - name
      - parentId
      - type
      type: object
    inline_response_200_1:
      example:
        uid: uid
      properties:
        uid:
          type: string
...